Bump spdlog/fmt versions
Required a workaround for formatting a bitfield, see https://github.com/gabime/spdlog/issues/1926

+++ src/JanusFtl.cpp
@@ -215,7 +215,7 @@ void JanusFtl::IncomingRtcp(janus_plugin_session* handle, janus_plugin_rtcp* pac
             handlePsfbRtcpPacket(handle, rtcpHeader);
             break;
         default:
-            spdlog::info("Got unknown RTCP packet! Type: {}", rtcpHeader->type);
+            spdlog::info("Got unknown RTCP packet! Type: {}", (uint16_t)rtcpHeader->type);
             break;
         }
